Refactor HatchEggScreen and related components for improved structure and readability
